python/2021/fms_app/requirements.txt
Django==3.2.5
djangorestframework
django-truncate
Pillow
mysqlclient
uuid